\name{bayessurv-package}
\alias{bayessurv-package}
\alias{bayessurv}
\docType{package}
\title{BayesSurv: a package to estimate mortality parameters from capture-recapture data}
\description{This package estimates mortality with covariates from capture-recapture/recovery data in a Bayesian framework when many individuals are of unknown age. It includes tools for data checking, model diagnostics and outputs such as life-tables and plots.}
\details{
\tabular{ll}{
Package: \tab bayessurv\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-10\cr
License: \tab GNU GPL\cr
LazyLoad: \tab yes\cr
}

This package estimates mortality parameters from capture-recapture data where there are individuals with unknown birth/death dates.
It is currently under development
}
\author{Fernando Colchero <colchero@mpg.demogr.de>;
Owen Jones <jones@mpg.demogr.de>;
Maren Rebke <rebke@mpg.demogr.de>
}
\references{Colchero and Clark 2011 Bayesian inference on age-specific survival from capture-recapture data for censored and truncated data. Journal of Animal Ecology. 
Colchero, Jones and Rebke. BayesSurv - a package to estimate mortality parameters from incomplete mark-recapture data with covariates. Unpublished Manuscript.
}
\keyword{ package }
\seealso{}
\examples{}
