% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSurv.R
\docType{package}
\name{BayesSurv}
\alias{BayesSurv}
\alias{BayesSurv-package}
\title{Fit the Bayesian Cox Model}
\usage{
BayesSurv(survObj, priorPara, initial, nIter = 1, seed = 123)
}
\arguments{
\item{survObj}{a list containing observed data from \code{n} subjects;
\code{t}, \code{di}, \code{X}. See details for more information}

\item{priorPara}{a list containing prior parameter values}

\item{initial}{a list containing prior parameters' initial values}

\item{nIter}{the number of iterations of the chain}

\item{seed}{random seed}
}
\value{
An object of class \code{BayesSurv} is saved as
\code{obj_BayesSurv.rda} in the output file, including the following components:
\itemize{
\item input - a list of all input parameters by the user
\item output - a list of the all output estimates:
\itemize{
\item "\code{gamma.p}" - a matrix with MCMC intermediate estimates of the indicator variables of regression coefficients.
\item "\code{beta.p}" - a matrix with MCMC intermediate estimates of the regression coefficients.
\item "\code{h.p}" - a matrix with MCMC intermediate estimates of the increments in the cumulative baseline hazard in each interval.
}
\item call - the matched call.
}
}
\description{
This is the main function to fit a Bayesian Cox model with graph-structured 
selection priors for sparse identification of high-dimensional covariates.
}
\details{
BayesSurv
}
\examples{

# Load the example dataset

}
