% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_MCMC.R
\name{func_MCMC}
\alias{func_MCMC}
\title{Function to run MCMC sampling (information below to be updated)}
\usage{
func_MCMC(survObj, priorPara, initial, num.reps, S, method, MRF_2b, seed)
}
\arguments{
\item{survObj}{a list containing observed data from \code{n} subjects;
\code{t}, \code{di}, \code{x}. See details for more information}
}
\value{
An object of ...
}
\description{
This a
}
\details{
BayesSurv
}
\examples{

# Load the example dataset


}
