\name{plot.marginal.bayesGspline}
\alias{plot.marginal.bayesGspline}
\title{
  Plot an object of class marginal.bayesGspline
}  
\description{
  This function plots an object created by \code{\link{marginal.bayesGspline}}.
}
\usage{
\method{plot}{marginal.bayesGspline}(x, type = "l", lty=1, bty = "n",
   xlab1, ylab1, main1, xlab2, ylab2, main2, sub, \dots)
}
\arguments{
  \item{x}{an object of class \code{marginal.bayesGspline}.}
  \item{type, lty, bty, xlab1, ylab1, main1, xlab2, ylab2, main2, sub, \dots}{other
     arguments passed to the \code{plot.default} function.}
}  
\author{
  \eqn{\mbox{Arno\v{s}t Kom\'arek}}{Arnost Komarek} \email{arnost.komarek@med.kuleuven.be}
}
\keyword{hplot}
