% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BayesSurv.R
\name{predict.BayesSurv}
\alias{predict.BayesSurv}
\title{Predict survival risk (Not yet subgroups)}
\usage{
\method{predict}{BayesSurv}(
  object,
  survObj.new,
  type = "brier",
  method = "mean",
  times = NULL,
  subgroup = 1,
  ...
)
}
\arguments{
\item{object}{fitted object obtained with \code{BayesSurv}}

\item{survObj.new}{a list containing observed data from new subjects with
components \code{t}, \code{di}, \code{x}. If \code{type} is among
\code{c("hazard", "cumhazard", "survival")}, only \code{survObj.new$X} is 
needed.}

\item{type}{option to chose for predicting brier scores with \code{type="brier"} 
or one of \code{type=c("brier", "hazard", "cumhazard", "survival")})}

\item{method}{option to use the posterior mean (\code{"mean"}) of coefficients
for prediction or Bayesian model averaging (\code{"BMA"}) for prediction}

\item{times}{time points at which to evaluate the risks. If \code{NULL}
(default), the event/censoring times are used. If \code{type="brier"}, the
largest one of the \code{times} is used}

\item{subgroup}{index of the subgroup in \code{survObj.new} for prediction. 
Default value is 1}

\item{\dots}{not used}
}
\description{
Predict survival probability, (cumulative) hazard or (integrated) Brier scores based on Cox regression models
}
\details{
BayesSurv
}
\examples{

library("survival")
library("GGally")
library("BayesSurv")

# Load the example dataset
data("simData", package = "BayesSurv")

dataset = list("X" = simData[[1]]$X, 
               "t" = simData[[1]]$time,
               "di" = simData[[1]]$status)

# Initial value: null model without covariates
log.like  = coxph( Surv(dataset$t, dataset$di, type = c('right')) ~ 1 )$loglik 
initial = list("gamma.ini" = rep(0, ncol(dataset$X)), 
               "beta.ini" = rep(0, ncol(dataset$X)), 
               "log.like.ini" = log.like)
# Prior parameters
priorParaPooled = list(
  #"eta0"   = eta0,                   # prior of baseline hazard
  #"kappa0" = kappa0,                 # prior of baseline hazard
  "c0"     = 2,                      # prior of baseline hazard
  "tau"    = 0.0375,                 # sd for coefficient prior
  "cb"     = 20,                     # sd for coefficient prior
  "pi.ga"  = 0.02, #0.5, ga.pi,      # prior variable selection probability for standard Cox models
  "nu0"    = 0.05,                   # hyperparameter in graphical model
  "nu1"    = 5,                      # hyperparameter in graphical model
  "lambda" = 3,                      # hyperparameter in graphical model
  "a"      = -4, #a0,                # hyperparameter in MRF prior
  "b"      = 0.1, #b0,               # hyperparameter in MRF prior
  "G"      = simData$G               # hyperparameter in MRF prior
)    

\donttest{
# run Bayesian Cox with graph-structured priors
fit = BayesSurv(Data=dataset, priorPara=priorParaPooled, 
                initial=initial, nIter=100, seed=123)
# predict survival probabilities of the train data
predict(fit, survObj.new=dataset)
}

}
\keyword{survival}
