\name{print.bayesDensity}
\alias{print.bayesDensity}
\title{
   Print a summary for the density estimate based on the Bayesian model.
}
\description{
   This function prints a~object created by \code{\link{bayesDensity}}.
}
\usage{
\method{print}{bayesDensity}(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{bayesDensity}.}
  \item{\dots}{this is here for a consistency with a generic function.}  
}
\author{
  Arnošt Komárek \email{arnost.komarek[AT]mff.cuni.cz}
}
\keyword{print}

