% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePara.R
\name{UpdateGamma}
\alias{UpdateGamma}
\title{Subfunctions to update parameters (information below to be updated)}
\usage{
UpdateGamma(sobj, priorPara, ini, S, method, MRF_2b)
}
\arguments{
\item{y}{y}
}
\value{
An object of ...
}
\description{
This a
}
\details{
BayesSurv
}
\examples{

# Load the example dataset


}
