% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateRPlee11.R
\name{UpdateRPlee11}
\alias{UpdateRPlee11}
\title{Update coefficients of the Bayesian Cox Lasso Model (information below to be updated)}
\usage{
UpdateRPlee11(sobj, priorPara, ini, S, method)
}
\arguments{
\item{sobj}{object of ...}
}
\value{
An object of ...
}
\description{
This a
}
\details{
BayesSurv
}
\examples{

# Load the example dataset


}
