% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesSurv.R
\name{plot.BayesSurv}
\alias{plot.BayesSurv}
\title{create a plot of estimated coefficients}
\usage{
\method{plot}{BayesSurv}(x, type = "mean", interval = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{BayesSurv} or a matrix. If \code{x}
is a matrix, use \code{BayesSurv:::plot.BayesSurv(x)}}

\item{type}{type of point estimates of regression coefficients. One of
\code{c("mean", "median")}. Default is \code{mean}}

\item{interval}{logical argument to show 95\% credible intervals. Default
is \code{TRUE}}

\item{...}{additional arguments sent to \code{ggplot2::geom_point()}}
}
\value{
ggplot object
}
\description{
Plot point estimates of regression coefficients and 95\% credible intervals
}
\examples{

library("survival")
library("GGally")
library("BayesSurv")

# Load the example dataset
data("simData", package = "BayesSurv")

dataset = list("X" = simData[[1]]$X, 
               "t" = simData[[1]]$time,
               "di" = simData[[1]]$status)

# Initial value: null model without covariates
log.like  = coxph( Surv(dataset$t, dataset$di, type = c('right')) ~ 1 )$loglik 
initial = list("gamma.ini" = rep(0, ncol(dataset$X)), 
               "beta.ini" = rep(0, ncol(dataset$X)), 
               "log.like.ini" = log.like)
               
# Prior parameters
priorParaPooled = list(
  #"eta0"   = eta0,                   # prior of baseline hazard
  #"kappa0" = kappa0,                 # prior of baseline hazard
  "c0"     = 2,                      # prior of baseline hazard
  "tau"    = 0.0375,                 # sd for coefficient prior
  "cb"     = 20,                     # sd for coefficient prior
  "pi.ga"  = 0.02, #0.5, ga.pi,      # prior variable selection probability for standard Cox models
  "nu0"    = 0.05,                   # hyperparameter in graphical model
  "nu1"    = 5,                      # hyperparameter in graphical model
  "lambda" = 3,                      # hyperparameter in graphical model
  "a"      = -4, #a0,                # hyperparameter in MRF prior
  "b"      = 0.1, #b0,               # hyperparameter in MRF prior
  "G"      = simData$G               # hyperparameter in MRF prior
)    

\donttest{
# run Bayesian Cox with graph-structured priors
fit = BayesSurv(survObj=dataset, priorPara=priorParaPooled, 
                initial=initial, nIter=100, seed=123)

# show posterior mean of coefficients and 95\% credible intervals
plot(fit) + 
  coord_flip() + 
    theme(axis.text.x = element_text(angle = 90, size = 7))
}

}
