% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateBH.R, R/updatePara.R
\name{UpdateGamma}
\alias{UpdateGamma}
\alias{UpdateBH}
\title{Update increment in cumulative baseline hazard}
\usage{
UpdateBH(sobj, priorPara, ini, S, method)

UpdateGamma(sobj, priorPara, ini, S, method, MRF_2b)
}
\arguments{
\item{sobj}{a list containing observed data}

\item{priorPara}{a list containing prior parameter values}

\item{ini}{a list containing prior parameters' initial values}

\item{S}{the number of subgroups}

\item{method}{a method option from 
\code{c("Pooled", "CoxBVSSL", "Sub-struct", "Subgroup")}}

\item{MRF_2b}{two different b in MRF prior for subgraphs G_ss and G_rs}
}
\value{
An object of ...

An object of ...
}
\description{
This contains subfunctions to update increment in cumulative baseline hazard

This contains subfunctions to update parameters gammas, betas, baseline 
hazard and graph learning parameters
}
\details{
BayesSurv

BayesSurv
}
\examples{

# Load the example dataset


# Load the example dataset

}
