% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateRPlee11.R
\name{UpdateRPlee11}
\alias{UpdateRPlee11}
\title{Update coefficients of the Bayesian Cox Lasso Model}
\usage{
UpdateRPlee11(sobj, priorPara, ini, S, method)
}
\arguments{
\item{sobj}{a list containing observed data}

\item{priorPara}{a list containing prior parameter values}

\item{ini}{a list containing prior parameters' initial values}

\item{S}{the number of subgroups}

\item{method}{a method option from 
\code{c("Pooled", "CoxBVSSL", "Sub-struct", "CoxBVSSL", "Sub-struct")}}
}
\value{
An object of ...
}
\description{
This an internal function to update coefficients of the Bayesian Cox Lasso Model
}
\details{
BayesSurv
}
\examples{

# Load the example dataset

}
